Reboot, using the Artists Theme instead of html5-bootstrap.
Jekyll produces HTML5 and CSS3 code from modular sources that let
me build a high quality website without distracting me with details
of the implementation; yet the Artists Theme is written with pedagogy
in mind, and makes it easier than Squarespace etc. to understand the
code driving the site. It is a great resource!